langchan-db-mcp/requirements.txt
langchain>=0.3.12
langchain-core>=0.2.0
langchain-community
langchain-ollama
python-dotenv
mysql-connector-python
